WIP Project structure refactoring
