+++ tools/single_experiment.py
@@ -8,7 +8,7 @@ from zeroshoteval.data.dataset import ObjEmbeddingDataset
 from zeroshoteval.evaluation.classification import classification_procedure
 from zeroshoteval.utils.defaults import default_setup
 from zeroshoteval.utils.parser import load_config, parse_args
-from zeroshoteval.models.build import build_zsl
+from zeroshoteval.zeroshotnets.build import build_zsl
 
 
 def setup(args: Namespace) -> CfgNode:
@@ -42,35 +42,36 @@ def experiment(cfg: CfgNode) -> None:
     # Extraction
     pass
 
-    # Embeddings saving to disk
+    # Embeddings saving to disk (save as torch tensor)
     pass
 
     # STEP 2 - MODEL TRAINING
-    # Pass extracted embeddings to:
-    # Option 1 - end-to-end ZSL + classification (or other) model to train it.
-    # Option 2 - to separate zero-shot net to train an embedding extractor.
+    # Pass extracted embeddings to a zero-shot neural network to train it as an
+    # embedding extractor.
     # --------------------------------------------------------------------------
 
     # Embeddings loading from disk
-    # TODO: write a script for transforming matfiles to folder with embeddings
+    # TODO: write a script for transforming matfiles to a saved torch tensor
     # TODO: merge Misha's changes of dataloader
     # TODO: Replace ObjEmbeddingDataset with loader from separate files
     data = ObjEmbeddingDataset(cfg.DATA.FEAT_EMB.PATH, ["IMG"], cfg.VERBOSE)
 
     # Training
     train_procedure = build_zsl(cfg)
-    zsl_emb_dataset, csl_train_indice, csl_test_indice = train_procedure(cfg, data)
+    zsl_emb_dataset, csl_train_indice, csl_test_indice = train_procedure(cfg)
 
     # STEP 3 - MODEL INFERENCE
-    # Apply trained model to test data and (similar to the prev step)
-    # Option 1 - evaluate model quality on one of tasks (classification, etc.)
-    # Option 2 - extract zero-shot embeddings.
+    # Apply trained model to test data and (similar to the prev step) extract
+    # zero-shot embeddings.
     # --------------------------------------------------------------------------
 
     # TODO: extract gen_synthetic_data from CADA-VAE and move it to inference section
     pass
 
-    # STEP 4 - MODEL EVALUATION (for zero-shot embeddings only)
+    # Zero-shot embeddings saving to disk (save as torch tensor)
+    pass
+
+    # STEP 4 - MODEL EVALUATION
     # Pass extracted zero-shot embeddings to one of evaluation tasks (
     # classification, clustering, verification, etc.)
     # --------------------------------------------------------------------------
